refactor: make it reusable for async plugins and other usages
