+++ packages/vue-markdown/package.json
@@ -52,7 +52,9 @@
   },
   "devDependencies": {
     "@types/hast": "^3.0.4",
+    "@types/mdast": "^4.0.4",
     "@types/unist": "^3.0.3",
+    "mdast": "^3.0.0",
     "mdast-util-to-hast": "^13.2.0",
     "unbuild": "^2.0.0"
   }
