+++ playground/package.json
@@ -18,6 +18,7 @@
     "vue": "^3.5.13"
   },
   "devDependencies": {
+    "@shikijs/rehype": "^3.6.0",
     "@vitejs/plugin-vue": "^5.2.1",
     "typescript": "~5.6.2",
     "vite": "^6.0.1",
